# server/api/swagger.yaml
info:
  version: 1.3.1
  title: Rosetta
  description: A Standard for Blockchain Interaction
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /network/status:
    post:
      summary: Get Network Status
      description: This endpoint returns the current status of the network requested.
        Any NetworkIdentifier returned by /network/list should be accessible
        here.
      operationId: networkStatus
      tags:
        - Network
      responses:
        "200":
          description: Expected response to a valid request
          schema:
            $ref: "#/definitions/NetworkStatusResponse"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"
      parameters:
        - required: true
          name: body
          in: body
          schema:
            $ref: "#/definitions/NetworkRequest"
      consumes:
        - application/json
      produces:
        - application/json
  /network/options:
    post:
      summary: Get Network Options
      description: This endpoint returns the version information and allowed
        network-specific types for a NetworkIdentifier. Any NetworkIdentifier
        returned by /network/list should be accessible here.  Because options
        are retrievable in the context of a NetworkIdentifier, it is possible to
        define unique options for each network.
      operationId: networkOptions
      tags:
        - Network
      responses:
        "200":
          description: Expected response to a valid request
          schema:
            $ref: "#/definitions/NetworkOptionsResponse"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"
      parameters:
        - required: true
          name: body
          in: body
          schema:
            $ref: "#/definitions/NetworkRequest"
      consumes:
        - application/json
      produces:
        - application/json
...
swagger: "2.0"
definitions:
  NetworkIdentifier:
    description: The network_identifier specifies which network a particular object
      is associated with.
    type: object
    required:
      - blockchain
      - network
    properties:
      blockchain:
        type: string
        example: bitcoin
      network:
        description: If a blockchain has a specific chain-id or network identifier, it
          should go in this field. It is up to the client to determine which
          network-specific identifier is mainnet or testnet.
        type: string
        example: mainnet
      sub_network_identifier:
        $ref: "#/definitions/SubNetworkIdentifier"
  SubNetworkIdentifier:
    description: In blockchains with sharded state, the SubNetworkIdentifier is
      required to query some object on a specific shard. This identifier is
      optional for all non-sharded blockchains.
    type: object
    required:
      - network
    properties:
      network:
        type: string
        example: shard 1
      metadata:
        type: object
        example:
          producer: "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
  BlockIdentifier:
    description: The block_identifier uniquely identifies a block in a particular network.
    type: object
    required:
      - index
      - hash
    properties:
      index:
        description: This is also known as the block height.
        type: integer
        format: int64
        example: 1123941
      hash:
        type: string
        example: "0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85"
...
  Peer:
    description: A Peer is a representation of a node's peer.
    type: object
    required:
      - peer_id
    properties:
      peer_id:
        type: string
        example: "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
      metadata:
        type: object
  Version:
    description: The Version object is utilized to inform the client of the versions
      of different components of the Rosetta implementation.
    type: object
    required:
      - rosetta_version
      - node_version
    properties:
      rosetta_version:
        description: The rosetta_version is the version of the Rosetta interface the
          implementation adheres to. This can be useful for clients looking to
          reliably parse responses.
        type: string
        example: 1.2.5
      node_version:
        description: The node_version is the canonical version of the node runtime. This
          can help clients manage deployments.
        type: string
        example: 1.0.2
      middleware_version:
        description: When a middleware server is used to adhere to the Rosetta
          interface, it should return its version here. This can help clients
          manage deployments.
        type: string
        example: 0.2.7
      metadata:
        description: Any other information that may be useful about versioning of
          dependent services should be returned here.
        type: object
  Allow:
    description: Allow specifies supported Operation status, Operation types, and
      all possible error statuses. This Allow object is used by clients to
      validate the correctness of a Rosetta Server implementation. It is
      expected that these clients will error if they receive some response that
      contains any of the above information that is not specified here.
    type: object
    required:
      - operation_statuses
      - operation_types
      - errors
    properties:
      operation_statuses:
        description: All Operation.Status this implementation supports. Any status that
          is returned during parsing that is not listed here will cause client
          validation to error.
        type: array
        items:
          $ref: "#/definitions/OperationStatus"
      operation_types:
        description: All Operation.Type this implementation supports. Any type that is
          returned during parsing that is not listed here will cause client
          validation to error.
        type: array
        items:
          type: string
          example: TRANSFER
      errors:
        description: All Errors that this implementation could return. Any error that is
          returned during parsing that is not listed here will cause client
          validation to error.
        type: array
        items:
          $ref: "#/definitions/Error"
  OperationStatus:
    description: OperationStatus is utilized to indicate which Operation status are
      considered successful.
    type: object
    required:
      - status
      - successful
    properties:
      status:
        description: The status is the network-specific status of the operation.
        type: string
      successful:
        description: An Operation is considered successful if the Operation.Amount
          should affect the Operation.Account. Some blockchains (like Bitcoin)
          only include successful operations in blocks but other blockchains
          (like Ethereum) include unsuccessful operations that incur a fee.  To
          reconcile the computed balance from the stream of Operations, it is
          critical to understand which Operation.Status indicate an Operation is
          successful and should affect an Account.
        type: boolean
    example:
      status: SUCCESS
      successful: true
  Timestamp:
    description: The timestamp of the block in milliseconds since the Unix Epoch.
      The timestamp is stored in milliseconds because some blockchains produce
      blocks more often than once a second.
    type: integer
    format: int64
    minimum: 0
    example: 1582833600000
...
  NetworkRequest:
    description: A NetworkRequest is utilized to retrieve some data specific
      exclusively to a NetworkIdentifier.
    type: object
    required:
      - network_identifier
    properties:
      network_identifier:
        $ref: "#/definitions/NetworkIdentifier"
      metadata:
        type: object
  NetworkStatusResponse:
    description: NetworkStatusResponse contains basic information about the node's
      view of a blockchain network.
    type: object
    required:
      - current_block_identifier
      - current_block_timestamp
      - genesis_block_identifier
      - peers
    properties:
      current_block_identifier:
        $ref: "#/definitions/BlockIdentifier"
      current_block_timestamp:
        $ref: "#/definitions/Timestamp"
      genesis_block_identifier:
        $ref: "#/definitions/BlockIdentifier"
      peers:
        type: array
        items:
          $ref: "#/definitions/Peer"
  NetworkOptionsResponse:
    description: NetworkOptionsResponse contains information about the versioning of
      the node and the allowed operation statuses, operation types, and errors.
    type: object
    required:
      - version
      - allow
    properties:
      version:
        $ref: "#/definitions/Version"
      allow:
        $ref: "#/definitions/Allow"
...
  Error:
    description: Instead of utilizing HTTP status codes to describe node errors
      (which often do not have a good analog), rich errors are returned using
      this object.
    type: object
    required:
      - code
      - message
      - retriable
    properties:
      code:
        description: Code is a network-specific error code. If desired, this code can be
          equivalent to an HTTP status code.
        type: integer
        format: int32
        minimum: 0
      message:
        description: Message is a network-specific error message.
        type: string
      retriable:
        description: An error is retriable if the same request may succeed if submitted
          again.
        type: boolean
...